fix compile error in skin editor sub-screen test case. also add missing karaoke skin config screen.
