+++ osu.Game.Rulesets.Karaoke.Tests/Screens/TestSceneConfigScreen.cs
@@ -2,11 +2,12 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using osu.Game.Rulesets.Karaoke.Screens.Skin.Config;
+using osu.Game.Rulesets.Karaoke.Skinning;
 
 namespace osu.Game.Rulesets.Karaoke.Tests.Screens
 {
     public class TestSceneConfigScreen : KaraokeSkinEditorScreenTestScene<ConfigScreen>
     {
-        protected override ConfigScreen CreateEditorScreen() => new();
+        protected override ConfigScreen CreateEditorScreen(KaraokeSkin karaokeSkin) => new(karaokeSkin);
     }
 }
