+++ osu.Game.Rulesets.Karaoke.Tests/Screens/TestSceneLayoutScreen.cs
@@ -2,25 +2,14 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using NUnit.Framework;
-using osu.Framework.Allocation;
 using osu.Game.Rulesets.Karaoke.Screens.Skin.Layout;
 using osu.Game.Rulesets.Karaoke.Skinning;
-using osu.Game.Skinning;
 
 namespace osu.Game.Rulesets.Karaoke.Tests.Screens
 {
     [TestFixture]
     public class TestSceneLayoutScreen : KaraokeSkinEditorScreenTestScene<LayoutScreen>
     {
-        protected override LayoutScreen CreateEditorScreen() => new();
-
-        [BackgroundDependencyLoader]
-        private void load(SkinManager skinManager)
-        {
-            // because skin is compared by id, so should change id to let skin manager info knows that skin has been changed.
-            var defaultSkin = DefaultKaraokeSkin.Default;
-            defaultSkin.ID = 100;
-            skinManager.CurrentSkinInfo.Value = defaultSkin;
-        }
+        protected override LayoutScreen CreateEditorScreen(KaraokeSkin karaokeSkin) => new(karaokeSkin);
     }
 }
